Typo in store controller when retail order info isn't valid. Modified javascript "order total" calculator, which is already kind of a mess, to do the right thing for non-integer-valued retail purchase amount.

+++ public/javascripts/application.js
@@ -130,7 +130,7 @@ function recalculate(target,selector,field_to_enable_if_nonzero,decplaces,no_pri
         }
         var qty   = (elt.tagName.match( /^select$/i ) ?
                  parseInt(elt.options[elt.selectedIndex].value)  :
-                 parseInt(elt.value)) ;
+                 parseFloat(elt.value)) ;
         if (isNaN(qty)) { qty = 0; }
         tot += (price * qty);
     }
